Breaking out Campground and Review Routes
- Did some refactoring with the campground and review routes with express router.

Off topic:
- Reverted the index.ejs page back to its orginal layout of 1 column.
