+++ views/campgrounds/index.ejs
@@ -1,55 +1,23 @@
 <% layout("/layouts/boilerplate") %>
 
-<h1 class="text-center">All Campgrounds</h1>
+<h1>All Campgrounds</h1>
 
-<div class="row">
-    <div class="col-6">
-        <% for (let i = 0; i < campgrounds.length/2; i++) { %>
-        <div class="card mb-3">
-            <div class="row">
-                <div class="col-md-6">
-                    <img class="img-fluid" alt="" src="<%= campgrounds[i].image %>">
-                </div>
-
-                <div class="col-md-6">
-                    <div class="card-body">
-                        <h5 class="card-title"><%= campgrounds[i].title %></h5>
-                        <p class="card-text"><%= campgrounds[i].description %></p>
-                        <p class="card-text">
-                            <small class="text-muted"><%= campgrounds[i].location %></small>
-                        </p>
-
-                        <a class="btn btn-primary" href="campgrounds/<%= campgrounds[i]._id %>">View
-                            <%= campgrounds[i].title %></a>
-                    </div>
-                </div>
+<% for (let campground of campgrounds) { %>
+<div class="card mb-3">
+    <div class="row">
+        <div class="col-md-4">
+            <img class="img-fluid" alt="" src="<%= campground.image %>">
+        </div>
+        <div class="col-md-8">
+            <div class="card-body">
+                <h5 class="card-title"><%= campground.title %></h5>
+                <p class="card-text"><%= campground.description %></p>
+                <p class="card-text">
+                    <small class="text-muted"><%= campground.location %></small>
+                </p>
+                <a class="btn btn-primary" href="campgrounds/<%= campground._id %>">View <%= campground.title %></a>
             </div>
         </div>
-        <% } %>
-    </div>
-
-    <div class="col-6">
-        <% for (let i = campgrounds.length - 1; i > campgrounds.length/2 ; i--) { %>
-        <div class="card mb-3">
-            <div class="row">
-                <div class="col-md-6">
-                    <img class="img-fluid" alt="" src="<%= campgrounds[i].image %>">
-                </div>
-
-                <div class="col-md-6">
-                    <div class="card-body">
-                        <h5 class="card-title"><%= campgrounds[i].title %></h5>
-                        <p class="card-text"><%= campgrounds[i].description %></p>
-                        <p class="card-text">
-                            <small class="text-muted"><%= campgrounds[i].location %></small>
-                        </p>
-
-                        <a class="btn btn-primary" href="campgrounds/<%= campgrounds[i]._id %>">View
-                            <%= campgrounds[i].title %></a>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <% } %>
     </div>
 </div>
+<% } %>
